refactor(inference): move portable MTP weight resolver
Fixes #838
